feat: ✨ add routers, clean up code
