+++ apps/api/src/routes/healthRouter.ts
@@ -1,0 +1,73 @@
+/*
+ * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import express, { type Request, type Response } from 'express';
+
+import { getHealth, Status } from '@/app-health.ts';
+import { serverConfig } from '@/config/serverConfig.ts';
+
+const healthRouter = express.Router();
+
+/**
+ * @openapi
+ * /health:
+ *  get:
+ *    tags:
+ *      - Health
+ *    summary: Get status of app and various app components.
+ *    responses:
+ *      '500':
+ *          description: 'App is not fully healthy'
+ *          content:
+ *            application/json:
+ *              schema:
+ *                $ref: '#/components/responses/RequestValidationError'
+ *      '200':
+ *          description: 'All components reporting healthy status'
+ *          content:
+ *              application/json:
+ *                  schema:
+ *                    type: object
+ *                    properties:
+ *                      version:
+ *                        type: string
+ *                      health:
+ *                        type: object
+ *                        additionalProperties:
+ *                         type: object
+ *                         properties:
+ *                             status:
+ *                                 type: string
+ *                             statusText:
+ *                                 type: string
+ */
+healthRouter.get('/health', (_req: Request, res: Response) => {
+	const health = getHealth();
+	const resBody = {
+		version: serverConfig.npm_package_version,
+		health,
+	};
+	if (health.all.status != Status.OK) {
+		res.status(500).json(resBody);
+		return;
+	}
+	res.status(200).json(resBody);
+});
+
+export default healthRouter;
